refactor: decompose LmsRepository and implement foreground transfer services
The previous implementation used a single `LmsRepository` for all operations, which became difficult to maintain. This overhaul splits the data layer into feature-specific repositories and migrates file transfer logic to dedicated foreground services for better reliability.

Key changes include:
- **Modular Repositories**: Split `LmsRepository` into `CourseRepository`, `AssignmentRepository`, `AttendanceRepository`, `MeetingRepository`, and `TransferRepository`.
- **Foreground Transfer Services**: Introduced `LmsDownloadService` and `LmsUploadService` to handle long-running file operations independently of the UI lifecycle.
- **Robust Error Handling**: Integrated `kotlin-result` across the data and ViewModel layers to replace standard `Result` or exception-based flows with functional error handling.
- **Storage Layer Overhaul**: Refactored `StorageDataSource` to handle `ParcelFileDescriptor` management and unique filename generation, facilitating safer IO between the JVM and the Rust native layer.
- **Native Interop Enhancements**: Updated the Rust `InternetDataSource` to support automatic re-authentication via a `CredentialProvider` interface and improved streaming using raw file descriptors.
- **Notification Throttling**: Updated `NotificationHelper` with a coroutine-based queue to throttle progress updates, preventing system UI overhead during high-frequency transfer events.
- **Data Type Standardization**: Migrated `meetingNumber` from `Byte` to `Int` across the database schema, models, and UI components for consistency.
- **UI & Lifecycle**: Simplified `MainActivity` by removing the splash screen API dependency and streamlining notification permission requests.

+++ app/src/main/kotlin/com/gaje48/lms/data/db/Entity.kt
@@ -41,7 +41,7 @@ data class CourseEntity(
 data class MeetingEntity(
     @PrimaryKey val meetingUrl: String,
     val courseCode: String,
-    val meetingNumber: Byte,
+    val meetingNumber: Int,
 )
 
 @Entity(
